fix(api): fixed nits
refs. metriport/metriport-internal#1350
